refactor: PageResponse common 패키지로 이동
- 필드명 변경 : programs -> contents
