+++ eeos/src/main/java/com/blackcompany/eeos/program/application/usecase/GetProgramsUsecase.java
@@ -1,6 +1,6 @@
 package com.blackcompany.eeos.program.application.usecase;
 
-import com.blackcompany.eeos.program.application.dto.PageResponse;
+import com.blackcompany.eeos.common.presentation.response.PageResponse;
 import com.blackcompany.eeos.program.application.dto.QueryProgramsResponse;
 
 public interface GetProgramsUsecase {
