+++ eeos/src/main/java/com/blackcompany/eeos/program/presentation/docs/GuestProgramApi.java
@@ -2,7 +2,7 @@ package com.blackcompany.eeos.program.presentation.docs;
 
 import com.blackcompany.eeos.common.presentation.response.ApiResponse;
 import com.blackcompany.eeos.common.presentation.response.ApiResponseBody.SuccessBody;
-import com.blackcompany.eeos.program.application.dto.PageResponse;
+import com.blackcompany.eeos.common.presentation.response.PageResponse;
 import com.blackcompany.eeos.program.application.dto.QueryProgramResponse;
 import com.blackcompany.eeos.program.application.dto.QueryProgramsResponse;
 import io.swagger.v3.oas.annotations.Operation;
